Tools to launch tmux quickly
The newmux command is a shortcut for quickly starting a new tmux session
with a name. It reconnects to an existing tmux session with the same
name if one exists.

The muxhere command is like newmux but uses the name of the present
working directory as the session name.

An example use of newmux is to run `newmux upgrade` when you want to run
commands to upgrade your OS. These commands have a chance of breaking X,
your terminal, etc. If you find yourself at a console with a login
prompt, you can run `newmux upgrade` again to reconnet to the existing
upgrade in progress.

The muxhere command is useful in the typical case of where you have to
work on a project specific to a directory. For example, to start working
on your project called `dog-petter`, you would `cd dog-petter` and then
run `muxhere`. It will connect to any existing tmux session for that
project if one exists, and otherwise it will start a new one.